Update installLisk.sh to accept an -f parameter to a local build - Closes #38

+++ downloaded/installLisk.sh
@@ -102,9 +102,16 @@ user_prompts() {
 }
 
 download_lisk() {
+	if [[ "$LOCAL_TAR" ]]; then
+		echo -e "\\nUsing local binary $LOCAL_TAR"
+		LISK_VERSION="$LOCAL_TAR"
+		LISK_DIR=${LISK_VERSION%.tar.gz}
+		return
+	fi
+
 	LISK_VERSION=lisk-$UNAME.tar.gz
 
-	LISK_DIR=$(echo "$LISK_VERSION" | cut -d'.' -f1)
+	LISK_DIR=${LISK_VERSION%.tar.gz}
 
 	rm -f "$LISK_VERSION" "$LISK_VERSION".SHA256 &> /dev/null
 
@@ -130,14 +137,17 @@ download_lisk() {
 }
 
 install_lisk() {
-	echo -e '\\nExtracting Lisk binaries to '"$LISK_INSTALL"
+	echo -e '\nExtracting Lisk binaries to '"$LISK_INSTALL"
 
 	tar -xzf "$LISK_VERSION" -C "$LISK_LOCATION"
 
 	mv "$LISK_LOCATION/$LISK_DIR" "$LISK_INSTALL"
 
-	echo -e "\\nCleaning up downloaded files"
-	rm -f "$LISK_VERSION" "$LISK_VERSION".SHA256
+	# if user is specifying a tarball, we probably don't want to delete it
+	if [[ ! $LOCAL_TAR ]]; then
+		echo -e "\\nCleaning up downloaded files"
+		rm -f "$LISK_VERSION" "$LISK_VERSION".SHA256
+	fi
 }
 
 configure_lisk() {
@@ -164,9 +174,11 @@ cleanup_installation() {
 
 	cd ../ || exit 2
 
-	echo -e "\\nRemoving Lisk directory and installation files"
-	rm -rf "$LISK_INSTALL"
-	rm -f "$LISK_VERSION" "$LISK_VERSION".SHA256
+	if [[ ! $LOCAL_TAR ]]; then
+		echo -e "\\nRemoving Lisk directory and installation files"
+		rm -rf "$LISK_INSTALL"
+		rm -f "$LISK_VERSION" "$LISK_VERSION".SHA256
+	fi
 
 	if [[ "$FRESH_INSTALL" == false ]]; then
 		echo -e "Restoring old Lisk installation"
@@ -269,10 +281,11 @@ usage() {
 
 parse_option() {
 	OPTIND=2
-	while getopts :d:r:u:h0: OPT; do
+	while getopts :d:f:r:u:h0: OPT; do
 		 # shellcheck disable=SC2220
 		 case "$OPT" in
 			 d) LISK_LOCATION="$OPTARG" ;;
+			 f) LOCAL_TAR="$OPTARG" ;;
 			 r) RELEASE="$OPTARG" ;;
 			 h) REBUILD=true ;;
 			 u) URL="$OPTARG" ;;
